feat: conftest tests
This was a huge mess: exposing the docker-registry was harder then expected due to security issues and containerd trust issues. In the end I did not manage to expose the docker-registry properly (I can push and pull from it, but conftest and oras do not manage to connect to it from within the container due to the containers not having access to localhost).

I think either exposing hostnetwork=true for the conftest-pull step or using let's encrypt for an actual cert might work.